Remove optionality of Loadout subscripts Closes #6
Technically it's possible for this to result in a
crash for a brand new character, or a character
that has not been played since a Destiny patch
where a weapon slot could be left empty due to
slot changes (e.g. 2.0.0). We should account for
these cases, but the best way to do so is to
simply trash the associated character object,
which could be unexpected to the user and should
probably be handled with care (read: do it later).
Else we must use Optionals or throw.